rename import script
